requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
trafilatura==1.6.0
sentence-transformers==2.2.2
huggingface-hub==0.16.4 
faiss-cpu==1.7.4
langchain==0.0.346
numpy==1.24.3
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
cachetools==5.3.2